Implement stream in findTask.
TaskList class: findTask is using for-loop
to see if any task name matches the keyword.

Change to stream for easier modification
reusability and testing.

The Arraylist tasks is converted to a stream, and
transformed using fiter, map and collect function
to get the output.

It is done this way to efficiently append the
tasks' name that match the keyword to the output.